Ported application to the new loader
