--- common/src/main/java/eu/unipv/epsilon/enigma/io/builderdefaults/RefBuilderDefaults.java
@@ -1,15 +1,0 @@
-package eu.unipv.epsilon.enigma.io.builderdefaults;
-
-import java.util.zip.ZipFile;
-
-public class RefBuilderDefaults implements BuilderDefaultsFactory<String> {
-
-    public DefaultFieldProvider getCollectionDefaults(ZipFile zipFile) {
-        return new QuestCollectionDefaults(zipFile);
-    }
-
-    public DefaultFieldProvider getQuestDefaults(int index) {
-        return new QuestDefaults(index);
-    }
-
-}
